new tilemap, and working world generation

+++ src/game/world/components.rs
@@ -4,5 +4,5 @@ use bevy::prelude::*;
 pub struct Tile {
     pub x: i32,
     pub y: i32,
-    pub sprite: i32,
+    pub sprite: usize,
 }
